Cleaning up some code
