+++ common/contract.ts
@@ -101,6 +101,5 @@ export const OUTCOME_TYPES = ['BINARY', 'FREE_RESPONSE', 'NUMERIC'] as const
 export const MAX_QUESTION_LENGTH = 480
 export const MAX_DESCRIPTION_LENGTH = 10000
 export const MAX_TAG_LENGTH = 60
-export const MAX_GROUP_PARTS_LENGTH = 100
 
 export const CPMM_MIN_POOL_QTY = 0.01
