+++ functions/src/backup-db.ts
@@ -50,7 +50,6 @@ export const backupDb = functions.pubsub
           'followers',
           'answers',
           'txns',
-          'messages',
         ],
       })
       .then((responses) => {
