+++ functions/src/index.ts
@@ -11,8 +11,6 @@ export * from './create-user'
 export * from './create-answer'
 export * from './on-create-bet'
 export * from './on-create-comment'
-export * from './on-group-follow'
-export * from './on-group-delete'
 export * from './on-view'
 export * from './unsubscribe'
 export * from './update-metrics'
