+++ web/pages/groups.tsx
@@ -10,7 +10,7 @@ import { Title } from 'web/components/title'
 import { UserLink } from 'web/components/user-page'
 import { useGroups, useMemberGroupIds } from 'web/hooks/use-group'
 import { useUser } from 'web/hooks/use-user'
-import { groupPath, listAllGroups, useMembers } from 'web/lib/firebase/groups'
+import { groupPath, listAllGroups } from 'web/lib/firebase/groups'
 import { getUser, User } from 'web/lib/firebase/users'
 import { Tabs } from 'web/components/layout/tabs'
 import { GroupMembersList } from 'web/pages/group/[...slugs]'
